Change semantics of index-batch-size
The old semantics is to request 'index-batch-size' number of blocks in
one RPC request.

This sometimes causes a very large peak memory usage in bitcoind, as
index-batch-size number of potentially 32MB blocks need to be serialized as
JSON in one request.

The new semantics is add a sync buffer of size 'index-batch-size' to the
electrum server and continuously try to fill the buffer by fetching one
block at a time.