Update configuration defaults
Change defaults to listen for RPC and websocket connections on all
intefaces.

The project we forked from aimed to be a private server, so listening
only on localhost made sense. However, ElectrsCash is aimed to be a
public facing server.

The blocktxid cache is reduced and tx cache is incresed the same amount.
Observations shown in production show that tx cache is more used.

Ports for testnet4 and scalenet have been added and the description
updated.